Update 1-2-assignment.md
The introduction movie on Bucket is a bit out of the blue to me, and very high on all sorts of terms that are unknown to beginners....
Link to bucket not working.
The point of installing the VCS in context of that phase in the assignment is unclear to me.
The part in 2.2 from creating a folder to the end I do not understand.
Do not add another digit. The hierarchy tree is already huge: module - assignment - step - task ... you would want a subtask under task.